+++ src/main/java/com/whatachad/app/model/response/FacilityResponseDto.java
@@ -1,6 +1,6 @@
 package com.whatachad.app.model.response;
 
-import com.whatachad.app.model.domain.Address;
+import com.whatachad.app.model.vo.Address;
 import com.whatachad.app.model.domain.Facility;
 import com.whatachad.app.type.FacilityType;
 import lombok.*;
