+++ src/main/java/com/whatachad/app/model/response/UserResponseDto.java
@@ -4,7 +4,6 @@ import com.whatachad.app.type.UserMetaType;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
@@ -40,9 +39,9 @@ public class UserResponseDto {
 	private Map<UserMetaType, String> meta;
 
 	@Schema(hidden = true)
-	private LocalDateTime createAt;
+	private LocalDateTime createdAt;
 
 	@Schema(hidden = true)
-	private LocalDateTime updateAt;
+	private LocalDateTime updatedAt;
 
 }
