+++ src/main/java/com/whatachad/app/util/MailUtil.java
@@ -9,13 +9,13 @@ import jakarta.mail.internet.MimeMessage;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.MimeMessageHelper;
 
-public class MailUtils {
+public class MailUtil {
 
     private JavaMailSender mailSender;
     private MimeMessage message;
     private MimeMessageHelper messageHelper;
 
-    public MailUtils(JavaMailSender mailSender) throws MessagingException {
+    public MailUtil(JavaMailSender mailSender) throws MessagingException {
         this.mailSender = mailSender;
         message = this.mailSender.createMimeMessage();
         messageHelper = new MimeMessageHelper(message, true, "UTF-8");
