+++ src/test/java/com/whatachad/app/controller/FacilityControllerTest.java
@@ -1,7 +1,7 @@
 package com.whatachad.app.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.whatachad.app.model.domain.Address;
+import com.whatachad.app.model.vo.Address;
 import com.whatachad.app.model.domain.Facility;
 import com.whatachad.app.model.dto.FacilityDto;
 import com.whatachad.app.model.request.CreateFacilityRequestDto;
